+++ src/bespokeasm/assembler/engine.py
@@ -29,7 +29,7 @@ class Assembler:
                 is_verbose: int,
                 include_paths: list[str],
             ):
-        self.source_file = source_file
+        self._source_file = source_file
         self._output_file = output_file
         self._config_file = config_file
         self._generate_binary = generate_binary
@@ -73,9 +73,9 @@ class Assembler:
 
             # add its label to the global scope
         # find base file containing directory
-        include_dirs = set([os.path.dirname(self.source_file)]+list(self._include_paths))
+        include_dirs = set([os.path.dirname(self._source_file)]+list(self._include_paths))
 
-        asm_file = AssemblyFile(self.source_file, global_label_scope)
+        asm_file = AssemblyFile(self._source_file, global_label_scope)
         line_obs: list[LineObject] = asm_file.load_line_objects(
             self._model,
             include_dirs,
@@ -155,7 +155,12 @@ class Assembler:
             print('NOT writing byte code to binary image.')
 
         if self._enable_pretty_print:
-            pprinter = PrettyPrinterFactory.getPrettyPrinter(self._pretty_print_format, line_obs, self._model)
+            pprinter = PrettyPrinterFactory.getPrettyPrinter(
+                self._pretty_print_format,
+                line_obs,
+                self._model,
+                self._source_file,
+            )
             pretty_str = pprinter.pretty_print()
             if self._pretty_print_output == 'stdout':
                 print(pretty_str)